Repository: fixed bug; pri remove se nevolaji udalosti kdyz entita neni v repositari

+++ Orm/Repository/Repository.php
@@ -343,19 +343,22 @@ abstract class Repository extends Object implements IRepository
 		$entity = $entity instanceof IEntity ? $entity : $this->getById($entity);
 		$this->checkAttachableEntity(get_class($entity), $entity);
 
-		$this->events->fireEvent(Events::REMOVE_BEFORE, $entity);
-		if (isset($entity->id))
+		if (isset($entity->id) OR $entity->getRepository(false))
 		{
-			if ($this->getMapper()->remove($entity))
+			$this->events->fireEvent(Events::REMOVE_BEFORE, $entity);
+			if (isset($entity->id))
 			{
-				$this->entities[$entity->id] = false;
+				if ($this->getMapper()->remove($entity))
+				{
+					$this->entities[$entity->id] = false;
+				}
+				else
+				{
+					throw new BadReturnException(array($this->getMapper(), 'remove', 'TRUE', NULL, '; something wrong with mapper'));
+				}
 			}
-			else
-			{
-				throw new BadReturnException(array($this->getMapper(), 'remove', 'TRUE', NULL, '; something wrong with mapper'));
-			}
+			$this->events->fireEvent(Events::REMOVE_AFTER, $entity);
 		}
-		$this->events->fireEvent(Events::REMOVE_AFTER, $entity);
 		return true;
 	}
